% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alation.R
\name{getLatestResult}
\alias{getLatestResult}
\title{Get latest result from Alation}
\usage{
getLatestResult(id)
}
\arguments{
\item{id}{Alation Query ID for which you want the most recent execution}
}
\value{
The result as a data frame
}
\description{
Get latest result from Alation
}
\examples{
\dontrun{
df <- getLatestResult(123456)
}

}
