% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alation.R
\name{getLatestResultID}
\alias{getLatestResultID}
\title{Get latest result id from Alation}
\usage{
getLatestResultID(id)
}
\arguments{
\item{id}{Alation Query ID for which you want the most recent execution}
}
\value{
The result id
}
\description{
Get latest result id from Alation
}
\examples{
\dontrun{
result_id <- getLatestResultID(123456)
}

}
