% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alation.R
\name{getQuery}
\alias{getQuery}
\title{Get query from Alation}
\usage{
getQuery(id)
}
\arguments{
\item{id}{- the Alation Query ID}
}
\value{
The query as a character object
}
\description{
Get query from Alation
}
\examples{
\dontrun{
sql <- getQuery(123456)
}
}

