% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alation.R
\name{getResult}
\alias{getResult}
\title{Get result from Alation}
\usage{
getResult(id)
}
\arguments{
\item{id}{- the Alation Result ID}
}
\value{
The result as a data frame
}
\description{
Get result from Alation
}
\examples{
\dontrun{
df <- getResult(12345)
}
}

