% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alation.R
\name{storeToken}
\alias{storeToken}
\title{Get a token to use API}
\usage{
storeToken(token = "", url = "")
}
\arguments{
\item{url}{Alation URL .}

\item{user}{Alation username.}

\item{password}{Alation password.}
}
\value{
An Alation token used in subsequent calls.
}
\description{
As a side effect this caches the token for other API calls to utilise
}
\examples{
\dontrun{
storeToken(token="aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa", url="https://alation.corp.ebay.com")
}

}
